Final version of kill script (I hope)wq

+++ kill.sh
@@ -1,6 +1,12 @@
 #!/bin/bash
 pid="$(ps aux | grep cut | grep -vw "grep" | awk '{print $2}')"
 
-echo "Killing CUT, PID: $pid"
-kill $pid
+if [ "$pid" == "" ]
+then
+	echo "Did not find CUT running..."
+	exit 0
+else
 
+	echo "Killing CUT, PID: $pid"
+	kill $pid
+fi
